requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
langchain-community>=0.1.0
openrouter>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
streamlit-chat>=0.1.1
apify-client>=1.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
